+++ example-cli.py
@@ -16,10 +16,14 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from pypicolcd import PicoLCD
+from pypicolcd import font_meta
+from pypicolcd import get_font_meta
+from pypicolcd import find_resource
 from datetime import datetime
 import random
 import timeit
 from timeit import default_timer as best_timer
+import sys
 
 p = PicoLCD()
 if p.dc is None:
@@ -37,37 +41,52 @@ if p.dc is None:
 # you pass force_refresh_enable=True to set_pixel)
 
 p.clear()
+print("* cleared")
 p.verbose_enable = True
+print("* p.verbose_enable is " + str(p.verbose_enable))
 p.set_pixel((0, 0), True)
 x, y = 0, 0
-# p.draw_image((x, y), "images/maze.png")
-# p.draw_image((x, y), "images/qbist.png", brightness=20)
-# p.draw_image((x, y), "images/gradient-vertical.png", threshold=float(30./255.))
-# p.draw_image((x, y), "images/gradient-vertical.png", brightness=1)
-p.draw_image((x, y), "images/checker-gradient.png", brightness=1)
+# p.draw_image((x, y), find_resource("images/maze.png"))
+# p.draw_image((x, y), find_resource("images/qbist.png"), brightness=20)
+# p.draw_image((x, y), find_resource("images/gradient-vertical.png"), threshold=float(30./255.))
+# p.draw_image((x, y), find_resource("images/gradient-vertical.png"), brightness=1)
+p.draw_image((x, y), find_resource("images/checker-gradient.png"), brightness=1)
 # NOTE: draw_text uses row,col format which is y,x order
-x, y = 0, 0
-p_dfs = p.default_font_size
-p.draw_text(y, x,
-            "Default font is " + str(p_dfs) + "pt ninepin")
+x = 0
 
-x, y = 0, 1
-# p.draw_text(y, x, "Zephyrean works if tiny ",
-            # font_path="fonts/zephyreg.ttf",  # reg is illegible small
-            # font_size=7, threshold=.5)
-p.draw_text(y, x, "\"Zephyrean\" font: smallest 8pt readable font",
-            font_path="fonts/zephyrea.ttf",
-            font_size=8, threshold=.5)
-x, y = 0, 2
-p.draw_text(y, x, "\"Flottflott\" font: unreadable below font size 11",
-            font_path="fonts/Flottflott.ttf",
-            font_size=11, threshold=.5, erase_behind_enable=True)
-x, y = 0, 3
-p.draw_text(y, x, "\"Press Start\" font",
-            font_path="fonts/prstartk.ttf", font_size=6,
-            erase_behind_enable=True)
-print("Your LCD should now show info about included fonts"
-      " which are available on graphics type picoLCD models.")
+try_fonts = [None, "Zephyrean", "Flottflott", "Press Start"]
+p_dfs = p.default_font_size
+p_df = p.default_font
+for font_i in range(min(len(try_fonts), 4)):
+    name = try_fonts[font_i]
+    msg = None
+    meta = None
+    threshold = .5
+    erase_behind_enable = False
+    if name is None:
+        name = p_df
+        threshold =  None
+        meta = get_font_meta(name)
+        msg = "The default font is {}pt {}.".format(p_dfs, name.title())
+    else:
+        meta = get_font_meta(name)
+        msg = "\"{}\": {}".format(name, meta["note"])
+    print("* showing {}...".format(name))
+    if name == "Flottflott":
+        erase_behind_enable = True
+        pass
+    if threshold is not None:
+        p.draw_text(y, x, msg, font_path=meta["path"],
+                    font_size=meta["default_size"], threshold=.5,
+                    erase_behind_enable=erase_behind_enable)
+    else:
+        p.draw_text(y, x, msg, font_path=meta["path"],
+                    font_size=meta["default_size"],
+                    erase_behind_enable=erase_behind_enable)
+    y += 1
+
+print("* Your LCD should now show info about included fonts\n"
+      "  which are available on graphics type picoLCD models.")
 x, y = 211, 1
 outline = ((x, y), (x+45, y+13))
 p.draw_rect(outline, True, filled=False)
@@ -81,7 +100,7 @@ im_h = 8
 r = x + im_w
 b = y + im_h
 x, y = 128, 64-im_h
-max_count = 1000
+max_count = 120
 _fps_last_frame_tick = None
 _fps_accumulated_time = 0.
 _fps_accumulated_count = 0
@@ -89,8 +108,10 @@ last_update_s = None
 fps = None
 
 count = 0
-print("Press Ctrl C to exit. Drawing "
+print("* Press Ctrl C to exit. Drawing "
       + str(max_count) + " animation frames(s)...")
+prev_done = None
+sys.stderr.write("\n")
 while count < max_count:
     # p.draw_rect(((x, y), (r, b)), False, filled=False)
         # erase old copy (slow, lots of flicker)
@@ -108,7 +129,12 @@ while count < max_count:
     elif y + y_vel < 0:
         y_vel = -y_vel
     count += 1
-    print(str(count))
+    done = float(count) / float(max_count)
+    # print(str(count))
+    # if (prev_done is None) or (done - prev_done > .1):
+    sys.stderr.write("\r{:.0f}% ({} of {}): The square should be drawing over the LCD screen.".format(done * 100.0, count, max_count))
+    prev_done = done
+sys.stderr.write("\n")
 
 count = 0
 p.clear()
@@ -122,5 +148,6 @@ while count < max_count:
                             # set_fps_interval, otherwise None
     # fps = p.get_fps()  # always gets fps if has ever been calculated
     if fps is not None:
-        print("[ example.py ] fps: " + str(fps))
+        # print("[ example.py ] fps: " + str(fps))
+        p.draw_text(0, 0, "fps: " + str(fps), erase_behind_enable=true)