+++ CHANGELOG.md
@@ -13,3 +13,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Refactor the project to make it into a setuptools package.
   - Adjust install instructions to match (easier and using virtualenv).
 - Change case of classes from PicoLcd to PicoLCD
+
+
+## [git] - 2018-01-26
+### Added
+- make example.py [later renamed to example-cli.ty]
+- draw text
+- draw image with threshold OR dithering
+
+### Changed
+- (fix use of refresh_enable) make force_refresh_enable=True work where on=False